Finished Creating The gamelogic
The game logic is correct and works as intended so long as the user inputs a valid string(Rock, Papers, Scissors)
This "assignment" was not difficult but man, the syntax in this language is a lot more difficult to remember compared to python.
Nonetheless, I will learn this language and improve with it. I have the logic, now I just need to implement it when necessary.